chore(downsample_mask): add downsample_mask filter class
The filter takes a path to a PNG image (will be converted to grayscale) and expected resolution/channel count of the mask.

The grayscale mask is then dithered into a black/white mask of the same dimensions, and an `excluded(NebulaPoint const&)` function is provided to test whether a point shall be excluded according to the mask.

For debug purposes, the dithered mask is written to the same directory as the input mask, with the file ending changed to `_dithered.png`. If this fails, the filter will not throw but log a warning.

Signed-off-by: Max SCHMELLER <max.schmeller@tier4.jp>